# docs/swagger.yaml
definitions:
...
  api.ErrorResponse:
    properties:
      errors:
        items:
          type: string
        type: array
    type: object
...
  core.AlertBrief:
    properties:
      alert_id:
        type: string
      rule_id:
        type: string
      rule_name:
        type: string
      severity:
        type: string
      status:
        type: string
      timestamp:
        type: string
    type: object
...
  core.AlertLink:
    properties:
      alert_id:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      id:
        type: string
      link_type:
        description: e.g., "related", "duplicate", "escalation"
        type: string
      linked_alert_id:
        type: string
    type: object
  core.AlertLinkRequest:
    properties:
      description:
        type: string
      link_type:
        type: string
      linked_alert_id:
        type: string
    type: object
  core.AlertLinkResponse:
    properties:
      alert_id:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      link_type:
        type: string
      linked_alert:
        $ref: '#/definitions/core.AlertBrief'
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/alerts/{id}/related:
    get:
      description: Get all alerts that are linked to the specified alert
      parameters:
      - description: Alert ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/core.AlertLinkResponse'
            type: array
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: List related alerts
      tags:
      - alerts
    post:
      consumes:
      - application/json
      description: Create a bi-directional relationship between two alerts. This operation
        is idempotent - if the link already exists, returns 200 OK with the existing
        link.
      parameters:
      - description: Alert ID
        in: path
        name: id
        required: true
        type: string
      - description: Link request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/core.AlertLinkRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Link already exists (idempotent)
          schema:
            $ref: '#/definitions/core.AlertLink'
        "201":
          description: Link created
          schema:
            $ref: '#/definitions/core.AlertLink'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Alert not found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Link alerts
      tags:
      - alerts
  /api/v1/alerts/{id}/related/{related_id}:
    delete:
      description: Remove the bi-directional relationship between two alerts. This
        operation is idempotent - if the link doesn't exist, returns 204 No Content.
      parameters:
      - description: Alert ID
        in: path
        name: id
        required: true
        type: string
      - description: Related Alert ID to unlink
        in: path
        name: related_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No content (link deleted or didn't exist)
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Unlink alerts
      tags:
      - alerts
...
swagger: "2.0"